OC10548 - Altera tela de geração do siope.

+++ con4_siope.RPC.php
@@ -32,29 +32,47 @@ switch ($oParam->exec) {
 
         try {
 
-            $siope = new Siope;
-            $siope->setAno($iAnoUsu);
-            $siope->setInstit($iInstit);
-            $siope->setBimestre($iBimestre);
-            $siope->setPeriodo();
-            $siope->setFiltros();
-            $siope->setDespesasOrcadas();
-            $siope->setDespesas();
-            $siope->agrupaDespesas();
-            $siope->geraLinhaVazia();
-            $siope->ordenaDespesas();
-            $siope->setNomeArquivo($sNomeArq);
-            $siope->gerarSiope();
+            if (count($oParam->arquivos) > 0) {
 
-            if ($siope->getErroSQL() > 0 ) {
-                throw new Exception ("Ocorreu um erro ao gerar IC ".$siope->getErroSQL());
+                foreach ($oParam->arquivos as $sArquivo) {
+
+                    if ($sArquivo == 'despesa') {
+
+                        $siope = new Siope;
+                        $siope->setAno($iAnoUsu);
+                        $siope->setInstit($iInstit);
+                        $siope->setBimestre($iBimestre);
+                        $siope->setPeriodo();
+                        $siope->setFiltros();
+                        $siope->setDespesasOrcadas();
+                        $siope->setDespesas();
+                        $siope->agrupaDespesas();
+                        $siope->geraLinhaVazia();
+                        $siope->ordenaDespesas();
+                        $siope->setNomeArquivo($sNomeArq);
+                        $siope->gerarSiope();
+
+                        if ($siope->getErroSQL() > 0 ) {
+                            throw new Exception ("Ocorreu um erro ao gerar IC ".$siope->getErroSQL());
+                        }
+
+                        $oRetorno->caminho = $oRetorno->nome = "{$siope->getNomeArquivo()}.csv";
+
+                        system("rm -f {$siope->getNomeArquivo()}.zip");
+                        system("bin/zip -q {$siope->getNomeArquivo()}.zip $oRetorno->caminho");
+                        $oRetorno->caminhoZip = $oRetorno->nomeZip = "{$siope->getNomeArquivo()}.zip";
+
+                    }
+
+                    if ($sArquivo == 'receita') {
+
+
+
+                    }
+
+                }
+
             }
-
-            $oRetorno->caminho = $oRetorno->nome = "{$siope->getNomeArquivo()}.csv";
-
-            system("rm -f {$siope->getNomeArquivo()}.zip");
-            system("bin/zip -q {$siope->getNomeArquivo()}.zip $oRetorno->caminho");
-            $oRetorno->caminhoZip = $oRetorno->nomeZip = "{$siope->getNomeArquivo()}.zip";
 
         } catch(Exception $eErro) {
